Cleanup include. Move more to patterns.h
